Auth config with token

+++ app/Http/Controllers/CustomerController.php
@@ -32,7 +32,7 @@ class CustomerController extends Controller
             'firstname' => ['required', 'alpha'],
             'lastname' => ['required', 'alpha'],
             'email' => ['required', 'email', 'unique:customers,email'],
-            'phone' => ['required', 'numeric', 'digits:10', 'regex:/((98)|(97))(\d){8}/'],
+            'phone' => ['required', 'integer', 'digits:10', 'regex:/((98)|(97))(\d){8}/'],
             'address' => ['required'],
             'password' => ['required', 'min:8'],
             'citizenship_number' => ['required', 'integer',  'gt:0', 'digits:10', 'regex:/(\d){10}/'],
@@ -72,7 +72,7 @@ class CustomerController extends Controller
         $data = $request->validate([
             'firstname' => ['required', 'alpha'],
             'lastname' => ['required', 'alpha'],
-            'phone' => ['required', 'numeric', 'digits:10', 'regex:/((98)|(97))(\d){8}/'],
+            'phone' => ['required', 'integer', 'digits:10', 'regex:/((98)|(97))(\d){8}/'],
             'address' => ['required'],
             'citizenship_number' => ['required', 'integer',  'gt:0', 'digits:10', 'regex:/(\d){10}/'],
         ]);
